Getting acceptEx and disconnectEx working for WindowsServer. This will allow the server to accept new connections asynchronously instead of blocking. It also allows us to reuse sockets for new clients after the previous client process completes.
This implementation is super messy, but it works. I tested it by sending
10,000 requests 1ms apart for the index page. It was able to process all
of them without hitching. Therefore, even though I'm not happy with the
code quality right now, I want to create a commit at a known correct
point-in-time. I'll clean everything up in a future commit.

Next up is implementing matching behavior for the Linux server.